tech-studio-projects/other-projects/enterprise-it-network-simulation/java/src/main/java/com/system/simulation/devices/Router.java
package com.system.simulation.devices;

import com.system.simulation.network.config.NetworkConfig;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import java.util.HashMap;
import java.util.Map;

/**
 * Represents a network router with basic routing table management and packet handling capabilities.
 * The Router class can initialize, add, and remove routes, simulate attacks, and receive packets.
 */
public class Router {

    private static final Logger logger = LoggerFactory.getLogger(Router.class);

    private String routerName;                // Name of the router
    private String ipAddress;                 // IP address assigned to the router
    private Map<String, String> routingTable; // Routing table with destination-next hop mappings

    /**
     * Constructor for initializing Router with specific parameters.
     *
     * @param routerName   the name of the router.
     * @param ipAddress    the IP address of the router.
     * @param routingTable the initial routing table (destination-next hop mappings).
     */
    public Router(String routerName, String ipAddress, Map<String, String> routingTable) {
        this.routerName = routerName;
        this.ipAddress = ipAddress;
        this.routingTable = new HashMap<>(routingTable);
    }

    /**
     * Constructor for initializing Router using a NetworkConfig object.
     * Loads configurations like router name, IP address, and initial routing table from NetworkConfig.
     *
     * @param networkConfig configuration object containing router setup information.
     */
    public Router(NetworkConfig networkConfig) {
        this.routerName = networkConfig.getRouterName();
        this.ipAddress = networkConfig.getIpAddress();
        this.routingTable = networkConfig.getRoutingTable();
    }

    public Router() {

    }

    /**
     * Initializes the router and logs its routing table for simulation.
     */
    public void initialize() {
        try {
            logger.info("Initializing router '{}' with IP address: {}", routerName, ipAddress);
            logger.info("Initial routing table: {}", routingTable);
            // Additional setup for routing protocols like OSPF or RIP could be implemented here.
        } catch (Exception e) {
            logger.error("Error initializing router '{}': {}", routerName, e.getMessage());
        }
    }

    /**
     * Adds a new route to the router's routing table.
     *
     * @param destination the destination network address.
     * @param nextHop     the next hop router for reaching the destination.
     */
    public void addRoute(String destination, String nextHop) {
        try {
            routingTable.put(destination, nextHop);
            logger.info("Added route to '{}' via next hop '{}'", destination, nextHop);
        } catch (Exception e) {
            logger.error("Error adding route to '{}': {}", destination, e.getMessage());
        }
    }

    /**
     * Removes an existing route from the routing table.
     *
     * @param destination the destination network address to be removed.
     */
    public void removeRoute(String destination) {
        try {
            routingTable.remove(destination);
            logger.info("Removed route to '{}'", destination);
        } catch (Exception e) {
            logger.error("Error removing route to '{}': {}", destination, e.getMessage());
        }
    }

    /**
     * Retrieves the IP address of the router.
     *
     * @return the IP address of the router.
     */
    public String getIpAddress() {
        return ipAddress;
    }

    /**
     * Retrieves the current routing table.
     *
     * @return a copy of the routing table as a Map.
     */
    public Map<String, String> getRoutingTable() {
        return new HashMap<>(routingTable);
    }

    /**
     * Simulates the reception of a network packet by the router.
     *
     * @param packetContent content of the packet received.
     * @return
     */
    public boolean receivePacket(String packetContent) {
        logger.info("Router '{}' received packet: {}", routerName, packetContent);
        // Packet processing logic can be added here.
        return false;
    }

    /**
     * Simulates a Smurf attack on the router by sending multiple ICMP packets to a broadcast address.
     */
    public void simulateSmurfFlood() {
        logger.warn("Simulating Smurf Flood attack on router '{}'", routerName);
        // Logic to simulate broadcast flooding with ICMP packets.
    }

    /**
     * Simulates an ARP Spoofing attack, where the router sends fake ARP replies to poison ARP caches.
     */
    public void simulateARPSpoofing() {
        logger.warn("Simulating ARP Spoofing attack on router '{}'", routerName);
        // Logic to simulate ARP spoofing.
    }

    public void receivePoisonedARP(String victimIP, String attackerMAC) {
    }

    public boolean isArpCachePoisoned() {
        return false;
    }

    public int getLoad() {
        return 0;
    }

    public String isCompromised() {
        return "";
    }

    public int getDroppedPackets() {
        return 0;
    }

    public String getNetworkName() {
        return "";
    }
}
